Update README.md
We can use this project for operations such as opening a connection to the oracle database and reading, writing, updating or deleting. All we have to do is add the dll file that is created after the project is compiled as a reference to the project we want, and copy the json format settings file to the main directory where the application is running.